+++ lib/features/auth/data/models/auth_model.dart
@@ -1,8 +1,7 @@
 import 'dart:convert';
 
-import 'package:flutter_vtv/features/auth/data/models/user_info_model.dart';
-
 import '../../domain/entities/auth_entity.dart';
+import 'user_info_model.dart';
 
 class AuthModel extends AuthEntity {
   const AuthModel({
@@ -43,5 +42,7 @@ class AuthModel extends AuthEntity {
 
   String toJson() => json.encode(toMap());
 
+  /// - [source] is Json String
   factory AuthModel.fromJson(String source) => AuthModel.fromMap(json.decode(source) as Map<String, dynamic>);
+  factory AuthModel.fromJsonMap(Map<String, dynamic> source) => AuthModel.fromMap(source);
 }