+++ lib/features/auth/presentation/bloc/auth_state.dart
@@ -1,3 +1,4 @@
+// ignore_for_file: public_member_api_docs, sort_constructors_first
 part of 'auth_cubit.dart';
 
 enum AuthStatus {
@@ -49,8 +50,23 @@ class AuthState extends Equatable {
 
   @override
   List<Object?> get props => [
+        code,
         status,
         auth,
         message,
       ];
+
+  AuthState copyWith({
+    AuthStatus? status,
+    AuthEntity? auth,
+    String? message,
+    int? code,
+  }) {
+    return AuthState._(
+      status: status ?? this.status,
+      auth: auth ?? this.auth,
+      message: message ?? this.message,
+      code: code ?? this.code,
+    );
+  }
 }