+++ lib/features/auth/presentation/components/logged_widget.dart
@@ -1,22 +1,22 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_vtv/features/auth/domain/entities/user_info_entity.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 
+import '../bloc/auth_cubit.dart';
 import 'user_app_bar.dart';
 
 class LoggedWidget extends StatelessWidget {
   const LoggedWidget({
     super.key,
-    required this.userInfo,
   });
 
- final UserInfoEntity userInfo;
 
   @override
   Widget build(BuildContext context) {
+    final auth = context.read<AuthCubit>().state.auth;
     return Scaffold(
       appBar: buildUserAppBar(context),
       body: Center(
-        child: Text('Login with ${userInfo.toString()}'),
+        child: Text('Login with $auth'),
       ),
     );
   }