+++ lib/features/auth/presentation/components/text_field_custom.dart
@@ -77,7 +77,7 @@ class TextFieldCustom extends StatelessWidget {
             prefixIcon: prefixIcon,
             hintText: hint,
             border: OutlineInputBorder(
-              borderRadius: BorderRadius.circular(8),
+              borderRadius: BorderRadius.circular(12),
             ),
             focusedBorder: OutlineInputBorder(
               borderRadius: BorderRadius.circular(8),