+++ lib/features/auth/presentation/pages/user_home.dart
@@ -1,13 +1,11 @@
-import 'dart:developer';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_vtv/core/components/custom_widgets.dart';
-import 'package:flutter_vtv/features/auth/presentation/components/not_logged_widget.dart';
 import 'package:go_router/go_router.dart';
 
+import '../../../../core/presentation/components/custom_widgets.dart';
 import '../bloc/auth_cubit.dart';
 import '../components/logged_widget.dart';
+import '../components/not_logged_widget.dart';
 
 class UserHome extends StatelessWidget {
   const UserHome({super.key});
@@ -19,7 +17,6 @@ class UserHome extends StatelessWidget {
     return BlocConsumer<AuthCubit, AuthState>(
       listener: (context, state) {
         if (state.status == AuthStatus.unauthenticated) {
-          log('code ${state.code} message ${state.message}');
           if (state.message != null) {
             ScaffoldMessenger.of(context).showSnackBar(
               SnackBar(
@@ -45,7 +42,7 @@ class UserHome extends StatelessWidget {
         if (state.status == AuthStatus.unauthenticated) {
           return const NotLoggedWidget();
         } else if (state.status == AuthStatus.authenticated) {
-          return LoggedWidget(userInfo: state.auth!.userInfo);
+          return const LoggedWidget();
         } else {
           return loadingWidget;
         }