+++ test/helpers/test_helper.mocks.dart
@@ -4,19 +4,22 @@
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:async' as _i8;
-import 'dart:convert' as _i16;
-import 'dart:typed_data' as _i17;
+import 'dart:convert' as _i17;
+import 'dart:typed_data' as _i19;
 
-import 'package:connectivity_plus/connectivity_plus.dart' as _i14;
+import 'package:connectivity_plus/connectivity_plus.dart' as _i15;
 import 'package:connectivity_plus_platform_interface/connectivity_plus_platform_interface.dart'
-    as _i15;
+    as _i16;
 import 'package:dartz/dartz.dart' as _i2;
 import 'package:flutter_secure_storage/flutter_secure_storage.dart' as _i4;
 import 'package:flutter_vtv/core/error/failures.dart' as _i9;
-import 'package:flutter_vtv/core/helpers/secure_storage_helper.dart' as _i13;
+import 'package:flutter_vtv/core/helpers/secure_storage_helper.dart' as _i14;
+import 'package:flutter_vtv/core/network/base_response.dart' as _i3;
 import 'package:flutter_vtv/features/auth/data/data_sources/auth_data_source.dart'
     as _i11;
-import 'package:flutter_vtv/features/auth/data/models/auth_model.dart' as _i3;
+import 'package:flutter_vtv/features/auth/data/models/auth_model.dart' as _i12;
+import 'package:flutter_vtv/features/auth/data/models/user_info_model.dart'
+    as _i13;
 import 'package:flutter_vtv/features/auth/domain/dto/register_params.dart'
     as _i10;
 import 'package:flutter_vtv/features/auth/domain/entities/auth_entity.dart'
@@ -25,7 +28,7 @@ import 'package:flutter_vtv/features/auth/domain/repositories/auth_repository.da
     as _i7;
 import 'package:http/http.dart' as _i6;
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:mockito/src/dummies.dart' as _i12;
+import 'package:mockito/src/dummies.dart' as _i18;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
@@ -50,8 +53,9 @@ class _FakeEither_0<L, R> extends _i1.SmartFake implements _i2.Either<L, R> {
         );
 }
 
-class _FakeAuthModel_1 extends _i1.SmartFake implements _i3.AuthModel {
-  _FakeAuthModel_1(
+class _FakeDataResponse_1<T> extends _i1.SmartFake
+    implements _i3.DataResponse<T> {
+  _FakeDataResponse_1(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -60,9 +64,20 @@ class _FakeAuthModel_1 extends _i1.SmartFake implements _i3.AuthModel {
         );
 }
 
-class _FakeFlutterSecureStorage_2 extends _i1.SmartFake
+class _FakeSuccessResponse_2 extends _i1.SmartFake
+    implements _i3.SuccessResponse {
+  _FakeSuccessResponse_2(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
+
+class _FakeFlutterSecureStorage_3 extends _i1.SmartFake
     implements _i4.FlutterSecureStorage {
-  _FakeFlutterSecureStorage_2(
+  _FakeFlutterSecureStorage_3(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -71,8 +86,8 @@ class _FakeFlutterSecureStorage_2 extends _i1.SmartFake
         );
 }
 
-class _FakeAuthEntity_3 extends _i1.SmartFake implements _i5.AuthEntity {
-  _FakeAuthEntity_3(
+class _FakeAuthEntity_4 extends _i1.SmartFake implements _i5.AuthEntity {
+  _FakeAuthEntity_4(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -81,8 +96,8 @@ class _FakeAuthEntity_3 extends _i1.SmartFake implements _i5.AuthEntity {
         );
 }
 
-class _FakeResponse_4 extends _i1.SmartFake implements _i6.Response {
-  _FakeResponse_4(
+class _FakeResponse_5 extends _i1.SmartFake implements _i6.Response {
+  _FakeResponse_5(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -91,9 +106,9 @@ class _FakeResponse_4 extends _i1.SmartFake implements _i6.Response {
         );
 }
 
-class _FakeStreamedResponse_5 extends _i1.SmartFake
+class _FakeStreamedResponse_6 extends _i1.SmartFake
     implements _i6.StreamedResponse {
-  _FakeStreamedResponse_5(
+  _FakeStreamedResponse_6(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -128,7 +143,7 @@ class MockAuthRepository extends _i1.Mock implements _i7.AuthRepository {
       ) as _i8.Future<_i2.Either<_i9.Failure, _i5.AuthEntity>>);
 
   @override
-  _i8.Future<_i2.Either<_i9.Failure, _i5.AuthEntity>>
+  _i8.Future<_i2.Either<_i3.ErrorResponse, _i3.DataResponse<_i5.AuthEntity>>>
       loginWithUsernameAndPassword(
     String? username,
     String? password,
@@ -141,9 +156,10 @@ class MockAuthRepository extends _i1.Mock implements _i7.AuthRepository {
                 password,
               ],
             ),
-            returnValue:
-                _i8.Future<_i2.Either<_i9.Failure, _i5.AuthEntity>>.value(
-                    _FakeEither_0<_i9.Failure, _i5.AuthEntity>(
+            returnValue: _i8.Future<
+                _i2.Either<_i3.ErrorResponse,
+                    _i3.DataResponse<_i5.AuthEntity>>>.value(_FakeEither_0<
+                _i3.ErrorResponse, _i3.DataResponse<_i5.AuthEntity>>(
               this,
               Invocation.method(
                 #loginWithUsernameAndPassword,
@@ -153,7 +169,8 @@ class MockAuthRepository extends _i1.Mock implements _i7.AuthRepository {
                 ],
               ),
             )),
-          ) as _i8.Future<_i2.Either<_i9.Failure, _i5.AuthEntity>>);
+          ) as _i8.Future<
+              _i2.Either<_i3.ErrorResponse, _i3.DataResponse<_i5.AuthEntity>>>);
 
   @override
   _i8.Future<_i2.Either<_i9.Failure, void>> cacheAuth(
@@ -174,21 +191,23 @@ class MockAuthRepository extends _i1.Mock implements _i7.AuthRepository {
       ) as _i8.Future<_i2.Either<_i9.Failure, void>>);
 
   @override
-  _i8.Future<_i2.Either<_i9.Failure, void>> logout(String? refreshToken) =>
+  _i8.Future<_i2.Either<_i3.ErrorResponse, _i3.SuccessResponse>> logout(
+          String? refreshToken) =>
       (super.noSuchMethod(
         Invocation.method(
           #logout,
           [refreshToken],
         ),
-        returnValue: _i8.Future<_i2.Either<_i9.Failure, void>>.value(
-            _FakeEither_0<_i9.Failure, void>(
+        returnValue: _i8
+            .Future<_i2.Either<_i3.ErrorResponse, _i3.SuccessResponse>>.value(
+            _FakeEither_0<_i3.ErrorResponse, _i3.SuccessResponse>(
           this,
           Invocation.method(
             #logout,
             [refreshToken],
           ),
         )),
-      ) as _i8.Future<_i2.Either<_i9.Failure, void>>);
+      ) as _i8.Future<_i2.Either<_i3.ErrorResponse, _i3.SuccessResponse>>);
 
   @override
   _i8.Future<_i2.Either<_i9.Failure, void>> deleteAuth() => (super.noSuchMethod(
@@ -207,58 +226,80 @@ class MockAuthRepository extends _i1.Mock implements _i7.AuthRepository {
       ) as _i8.Future<_i2.Either<_i9.Failure, void>>);
 
   @override
-  _i8.Future<_i2.Either<_i9.Failure, void>> register(
+  _i8.Future<_i2.Either<_i3.ErrorResponse, _i3.SuccessResponse>> register(
           _i10.RegisterParams? registerParams) =>
       (super.noSuchMethod(
         Invocation.method(
           #register,
           [registerParams],
         ),
-        returnValue: _i8.Future<_i2.Either<_i9.Failure, void>>.value(
-            _FakeEither_0<_i9.Failure, void>(
+        returnValue: _i8
+            .Future<_i2.Either<_i3.ErrorResponse, _i3.SuccessResponse>>.value(
+            _FakeEither_0<_i3.ErrorResponse, _i3.SuccessResponse>(
           this,
           Invocation.method(
             #register,
             [registerParams],
           ),
         )),
-      ) as _i8.Future<_i2.Either<_i9.Failure, void>>);
+      ) as _i8.Future<_i2.Either<_i3.ErrorResponse, _i3.SuccessResponse>>);
 
   @override
-  _i8.Future<_i2.Either<_i9.Failure, bool>> isValidToken(
+  _i8.Future<_i2.Either<_i9.Failure, bool>> isExpiredToken(
           String? accessToken) =>
       (super.noSuchMethod(
         Invocation.method(
-          #isTokenExpired,
+          #isExpiredToken,
           [accessToken],
         ),
         returnValue: _i8.Future<_i2.Either<_i9.Failure, bool>>.value(
             _FakeEither_0<_i9.Failure, bool>(
           this,
           Invocation.method(
-            #isTokenExpired,
+            #isExpiredToken,
             [accessToken],
           ),
         )),
       ) as _i8.Future<_i2.Either<_i9.Failure, bool>>);
 
   @override
-  _i8.Future<_i2.Either<_i9.Failure, String>> getAccessToken(
-          String? refreshToken) =>
+  _i8.Future<_i2.Either<_i3.ErrorResponse, _i3.DataResponse<String>>>
+      getNewAccessToken() => (super.noSuchMethod(
+            Invocation.method(
+              #getNewAccessToken,
+              [],
+            ),
+            returnValue: _i8.Future<
+                    _i2
+                    .Either<_i3.ErrorResponse, _i3.DataResponse<String>>>.value(
+                _FakeEither_0<_i3.ErrorResponse, _i3.DataResponse<String>>(
+              this,
+              Invocation.method(
+                #getNewAccessToken,
+                [],
+              ),
+            )),
+          ) as _i8
+              .Future<_i2.Either<_i3.ErrorResponse, _i3.DataResponse<String>>>);
+
+  @override
+  _i8.Future<_i2.Either<_i3.ErrorResponse, _i3.SuccessResponse>> sendCode(
+          String? username) =>
       (super.noSuchMethod(
         Invocation.method(
-          #getAccessToken,
-          [refreshToken],
+          #sendCode,
+          [username],
         ),
-        returnValue: _i8.Future<_i2.Either<_i9.Failure, String>>.value(
-            _FakeEither_0<_i9.Failure, String>(
+        returnValue: _i8
+            .Future<_i2.Either<_i3.ErrorResponse, _i3.SuccessResponse>>.value(
+            _FakeEither_0<_i3.ErrorResponse, _i3.SuccessResponse>(
           this,
           Invocation.method(
-            #getAccessToken,
-            [refreshToken],
+            #sendCode,
+            [username],
           ),
         )),
-      ) as _i8.Future<_i2.Either<_i9.Failure, String>>);
+      ) as _i8.Future<_i2.Either<_i3.ErrorResponse, _i3.SuccessResponse>>);
 }
 
 /// A class which mocks [AuthDataSource].
@@ -270,7 +311,7 @@ class MockAuthDataSource extends _i1.Mock implements _i11.AuthDataSource {
   }
 
   @override
-  _i8.Future<_i3.AuthModel> loginWithUsernameAndPassword(
+  _i8.Future<_i3.DataResponse<_i12.AuthModel>> loginWithUsernameAndPassword(
     String? username,
     String? password,
   ) =>
@@ -282,7 +323,8 @@ class MockAuthDataSource extends _i1.Mock implements _i11.AuthDataSource {
             password,
           ],
         ),
-        returnValue: _i8.Future<_i3.AuthModel>.value(_FakeAuthModel_1(
+        returnValue: _i8.Future<_i3.DataResponse<_i12.AuthModel>>.value(
+            _FakeDataResponse_1<_i12.AuthModel>(
           this,
           Invocation.method(
             #loginWithUsernameAndPassword,
@@ -292,52 +334,194 @@ class MockAuthDataSource extends _i1.Mock implements _i11.AuthDataSource {
             ],
           ),
         )),
-      ) as _i8.Future<_i3.AuthModel>);
+      ) as _i8.Future<_i3.DataResponse<_i12.AuthModel>>);
 
   @override
-  _i8.Future<void> register(_i10.RegisterParams? registerDTO) =>
+  _i8.Future<_i3.SuccessResponse> register(_i10.RegisterParams? registerDTO) =>
       (super.noSuchMethod(
         Invocation.method(
           #register,
           [registerDTO],
         ),
-        returnValue: _i8.Future<void>.value(),
-        returnValueForMissingStub: _i8.Future<void>.value(),
-      ) as _i8.Future<void>);
-
-  @override
-  _i8.Future<void> disableRefreshToken(String? refreshToken) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #disableRefreshToken,
-          [refreshToken],
-        ),
-        returnValue: _i8.Future<void>.value(),
-        returnValueForMissingStub: _i8.Future<void>.value(),
-      ) as _i8.Future<void>);
-
-  @override
-  _i8.Future<String> getAccessToken(String? refreshToken) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #getAccessToken,
-          [refreshToken],
-        ),
-        returnValue: _i8.Future<String>.value(_i12.dummyValue<String>(
+        returnValue:
+            _i8.Future<_i3.SuccessResponse>.value(_FakeSuccessResponse_2(
           this,
           Invocation.method(
-            #getAccessToken,
+            #register,
+            [registerDTO],
+          ),
+        )),
+      ) as _i8.Future<_i3.SuccessResponse>);
+
+  @override
+  _i8.Future<_i3.SuccessResponse> revokeRefreshToken(String? refreshToken) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #revokeRefreshToken,
+          [refreshToken],
+        ),
+        returnValue:
+            _i8.Future<_i3.SuccessResponse>.value(_FakeSuccessResponse_2(
+          this,
+          Invocation.method(
+            #revokeRefreshToken,
             [refreshToken],
           ),
         )),
-      ) as _i8.Future<String>);
+      ) as _i8.Future<_i3.SuccessResponse>);
+
+  @override
+  _i8.Future<_i3.DataResponse<String>> getNewAccessToken(
+          String? refreshToken) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getNewAccessToken,
+          [refreshToken],
+        ),
+        returnValue: _i8.Future<_i3.DataResponse<String>>.value(
+            _FakeDataResponse_1<String>(
+          this,
+          Invocation.method(
+            #getNewAccessToken,
+            [refreshToken],
+          ),
+        )),
+      ) as _i8.Future<_i3.DataResponse<String>>);
+
+  @override
+  _i8.Future<_i3.DataResponse<_i12.AuthModel>> getUserProfile(
+          {required String? accessToken}) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getUserProfile,
+          [],
+          {#accessToken: accessToken},
+        ),
+        returnValue: _i8.Future<_i3.DataResponse<_i12.AuthModel>>.value(
+            _FakeDataResponse_1<_i12.AuthModel>(
+          this,
+          Invocation.method(
+            #getUserProfile,
+            [],
+            {#accessToken: accessToken},
+          ),
+        )),
+      ) as _i8.Future<_i3.DataResponse<_i12.AuthModel>>);
+
+  @override
+  _i8.Future<_i3.SuccessResponse> editUserProfile({
+    required String? accessToken,
+    required _i13.UserInfoModel? newInfo,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #editUserProfile,
+          [],
+          {
+            #accessToken: accessToken,
+            #newInfo: newInfo,
+          },
+        ),
+        returnValue:
+            _i8.Future<_i3.SuccessResponse>.value(_FakeSuccessResponse_2(
+          this,
+          Invocation.method(
+            #editUserProfile,
+            [],
+            {
+              #accessToken: accessToken,
+              #newInfo: newInfo,
+            },
+          ),
+        )),
+      ) as _i8.Future<_i3.SuccessResponse>);
+
+  @override
+  _i8.Future<_i3.SuccessResponse> requestOtpForResetPassword(
+          String? username) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #requestOtpForResetPassword,
+          [username],
+        ),
+        returnValue:
+            _i8.Future<_i3.SuccessResponse>.value(_FakeSuccessResponse_2(
+          this,
+          Invocation.method(
+            #requestOtpForResetPassword,
+            [username],
+          ),
+        )),
+      ) as _i8.Future<_i3.SuccessResponse>);
+
+  @override
+  _i8.Future<_i3.SuccessResponse> resetPassword({
+    required String? username,
+    required String? otp,
+    required String? newPassword,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #resetPassword,
+          [],
+          {
+            #username: username,
+            #otp: otp,
+            #newPassword: newPassword,
+          },
+        ),
+        returnValue:
+            _i8.Future<_i3.SuccessResponse>.value(_FakeSuccessResponse_2(
+          this,
+          Invocation.method(
+            #resetPassword,
+            [],
+            {
+              #username: username,
+              #otp: otp,
+              #newPassword: newPassword,
+            },
+          ),
+        )),
+      ) as _i8.Future<_i3.SuccessResponse>);
+
+  @override
+  _i8.Future<_i3.SuccessResponse> changePassword({
+    required String? username,
+    required String? oldPassword,
+    required String? newPassword,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #changePassword,
+          [],
+          {
+            #username: username,
+            #oldPassword: oldPassword,
+            #newPassword: newPassword,
+          },
+        ),
+        returnValue:
+            _i8.Future<_i3.SuccessResponse>.value(_FakeSuccessResponse_2(
+          this,
+          Invocation.method(
+            #changePassword,
+            [],
+            {
+              #username: username,
+              #oldPassword: oldPassword,
+              #newPassword: newPassword,
+            },
+          ),
+        )),
+      ) as _i8.Future<_i3.SuccessResponse>);
 }
 
 /// A class which mocks [SecureStorageHelper].
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockSecureStorageHelper extends _i1.Mock
-    implements _i13.SecureStorageHelper {
+    implements _i14.SecureStorageHelper {
   MockSecureStorageHelper() {
     _i1.throwOnMissingStub(this);
   }
@@ -345,7 +529,7 @@ class MockSecureStorageHelper extends _i1.Mock
   @override
   _i4.FlutterSecureStorage get I => (super.noSuchMethod(
         Invocation.getter(#I),
-        returnValue: _FakeFlutterSecureStorage_2(
+        returnValue: _FakeFlutterSecureStorage_3(
           this,
           Invocation.getter(#I),
         ),
@@ -363,7 +547,7 @@ class MockSecureStorageHelper extends _i1.Mock
           #readAuth,
           [],
         ),
-        returnValue: _i8.Future<_i5.AuthEntity>.value(_FakeAuthEntity_3(
+        returnValue: _i8.Future<_i5.AuthEntity>.value(_FakeAuthEntity_4(
           this,
           Invocation.method(
             #readAuth,
@@ -406,28 +590,28 @@ class MockSecureStorageHelper extends _i1.Mock
 /// A class which mocks [Connectivity].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockConnectivity extends _i1.Mock implements _i14.Connectivity {
+class MockConnectivity extends _i1.Mock implements _i15.Connectivity {
   MockConnectivity() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i8.Stream<_i15.ConnectivityResult> get onConnectivityChanged =>
+  _i8.Stream<_i16.ConnectivityResult> get onConnectivityChanged =>
       (super.noSuchMethod(
         Invocation.getter(#onConnectivityChanged),
-        returnValue: _i8.Stream<_i15.ConnectivityResult>.empty(),
-      ) as _i8.Stream<_i15.ConnectivityResult>);
+        returnValue: _i8.Stream<_i16.ConnectivityResult>.empty(),
+      ) as _i8.Stream<_i16.ConnectivityResult>);
 
   @override
-  _i8.Future<_i15.ConnectivityResult> checkConnectivity() =>
+  _i8.Future<_i16.ConnectivityResult> checkConnectivity() =>
       (super.noSuchMethod(
         Invocation.method(
           #checkConnectivity,
           [],
         ),
-        returnValue: _i8.Future<_i15.ConnectivityResult>.value(
-            _i15.ConnectivityResult.bluetooth),
-      ) as _i8.Future<_i15.ConnectivityResult>);
+        returnValue: _i8.Future<_i16.ConnectivityResult>.value(
+            _i16.ConnectivityResult.bluetooth),
+      ) as _i8.Future<_i16.ConnectivityResult>);
 }
 
 /// A class which mocks [Client].
@@ -449,7 +633,7 @@ class MockHttpClient extends _i1.Mock implements _i6.Client {
           [url],
           {#headers: headers},
         ),
-        returnValue: _i8.Future<_i6.Response>.value(_FakeResponse_4(
+        returnValue: _i8.Future<_i6.Response>.value(_FakeResponse_5(
           this,
           Invocation.method(
             #head,
@@ -470,7 +654,7 @@ class MockHttpClient extends _i1.Mock implements _i6.Client {
           [url],
           {#headers: headers},
         ),
-        returnValue: _i8.Future<_i6.Response>.value(_FakeResponse_4(
+        returnValue: _i8.Future<_i6.Response>.value(_FakeResponse_5(
           this,
           Invocation.method(
             #get,
@@ -485,7 +669,7 @@ class MockHttpClient extends _i1.Mock implements _i6.Client {
     Uri? url, {
     Map<String, String>? headers,
     Object? body,
-    _i16.Encoding? encoding,
+    _i17.Encoding? encoding,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -497,7 +681,7 @@ class MockHttpClient extends _i1.Mock implements _i6.Client {
             #encoding: encoding,
           },
         ),
-        returnValue: _i8.Future<_i6.Response>.value(_FakeResponse_4(
+        returnValue: _i8.Future<_i6.Response>.value(_FakeResponse_5(
           this,
           Invocation.method(
             #post,
@@ -516,7 +700,7 @@ class MockHttpClient extends _i1.Mock implements _i6.Client {
     Uri? url, {
     Map<String, String>? headers,
     Object? body,
-    _i16.Encoding? encoding,
+    _i17.Encoding? encoding,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -528,7 +712,7 @@ class MockHttpClient extends _i1.Mock implements _i6.Client {
             #encoding: encoding,
           },
         ),
-        returnValue: _i8.Future<_i6.Response>.value(_FakeResponse_4(
+        returnValue: _i8.Future<_i6.Response>.value(_FakeResponse_5(
           this,
           Invocation.method(
             #put,
@@ -547,7 +731,7 @@ class MockHttpClient extends _i1.Mock implements _i6.Client {
     Uri? url, {
     Map<String, String>? headers,
     Object? body,
-    _i16.Encoding? encoding,
+    _i17.Encoding? encoding,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -559,7 +743,7 @@ class MockHttpClient extends _i1.Mock implements _i6.Client {
             #encoding: encoding,
           },
         ),
-        returnValue: _i8.Future<_i6.Response>.value(_FakeResponse_4(
+        returnValue: _i8.Future<_i6.Response>.value(_FakeResponse_5(
           this,
           Invocation.method(
             #patch,
@@ -578,7 +762,7 @@ class MockHttpClient extends _i1.Mock implements _i6.Client {
     Uri? url, {
     Map<String, String>? headers,
     Object? body,
-    _i16.Encoding? encoding,
+    _i17.Encoding? encoding,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -590,7 +774,7 @@ class MockHttpClient extends _i1.Mock implements _i6.Client {
             #encoding: encoding,
           },
         ),
-        returnValue: _i8.Future<_i6.Response>.value(_FakeResponse_4(
+        returnValue: _i8.Future<_i6.Response>.value(_FakeResponse_5(
           this,
           Invocation.method(
             #delete,
@@ -615,7 +799,7 @@ class MockHttpClient extends _i1.Mock implements _i6.Client {
           [url],
           {#headers: headers},
         ),
-        returnValue: _i8.Future<String>.value(_i12.dummyValue<String>(
+        returnValue: _i8.Future<String>.value(_i18.dummyValue<String>(
           this,
           Invocation.method(
             #read,
@@ -626,7 +810,7 @@ class MockHttpClient extends _i1.Mock implements _i6.Client {
       ) as _i8.Future<String>);
 
   @override
-  _i8.Future<_i17.Uint8List> readBytes(
+  _i8.Future<_i19.Uint8List> readBytes(
     Uri? url, {
     Map<String, String>? headers,
   }) =>
@@ -636,8 +820,8 @@ class MockHttpClient extends _i1.Mock implements _i6.Client {
           [url],
           {#headers: headers},
         ),
-        returnValue: _i8.Future<_i17.Uint8List>.value(_i17.Uint8List(0)),
-      ) as _i8.Future<_i17.Uint8List>);
+        returnValue: _i8.Future<_i19.Uint8List>.value(_i19.Uint8List(0)),
+      ) as _i8.Future<_i19.Uint8List>);
 
   @override
   _i8.Future<_i6.StreamedResponse> send(_i6.BaseRequest? request) =>
@@ -647,7 +831,7 @@ class MockHttpClient extends _i1.Mock implements _i6.Client {
           [request],
         ),
         returnValue:
-            _i8.Future<_i6.StreamedResponse>.value(_FakeStreamedResponse_5(
+            _i8.Future<_i6.StreamedResponse>.value(_FakeStreamedResponse_6(
           this,
           Invocation.method(
             #send,